Suppress testDiscoveryUsesNodesFromLastClusterState again
This test was reinstated in #95409 but in fact it sometimes fails, and
that is in fact the desired behaviour. With this commit we move it to
the suite of tests that intentionally do not work with atomic register
coordinators.
