+++ server/src/test/java/org/elasticsearch/cluster/coordination/CoordinatorVotingConfigurationTests.java
@@ -22,6 +22,7 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import static org.elasticsearch.cluster.coordination.AbstractCoordinatorTestCase.Cluster.DEFAULT_DELAY_VARIABILITY;
+import static org.elasticsearch.cluster.coordination.AbstractCoordinatorTestCase.Cluster.EXTREME_DELAY_VARIABILITY;
 import static org.elasticsearch.cluster.coordination.Coordinator.Mode.CANDIDATE;
 import static org.elasticsearch.cluster.coordination.ElectionSchedulerFactory.ELECTION_INITIAL_TIMEOUT_SETTING;
 import static org.elasticsearch.cluster.coordination.Reconfigurator.CLUSTER_AUTO_SHRINK_VOTING_CONFIGURATION;
@@ -486,6 +487,37 @@ public class CoordinatorVotingConfigurationTests extends AbstractCoordinatorTest
             cluster.clusterNodes.replaceAll(cn -> cn == restartingNode ? cn.restartedNode() : cn);
             cluster.stabilise();
             mockAppender.assertAllExpectationsMatched();
+        }
+    }
+
+    public void testDiscoveryUsesNodesFromLastClusterState() {
+
+        // This test only applies when using proper majority-based voting configurations. If we permit non-overlapping configurations then
+        // the isolated node can form a one-node cluster when it is healed, and the other nodes won't know be able to discover it without a
+        // proper discovery config since it wasn't in their last cluster states.
+        //
+        // (if we wanted this to work without a proper discovery config then we could e.g. record the master's publish address in the
+        // heartbeat blob, but we can assume that discovery is properly configured when using an atomic-register coordinator, and therefore
+        // there's no need for this extra complexity)
+
+        try (Cluster cluster = new Cluster(randomIntBetween(3, 5))) {
+            cluster.runRandomly();
+            cluster.stabilise();
+
+            final Cluster.ClusterNode partitionedNode = cluster.getAnyNode();
+            if (randomBoolean()) {
+                logger.info("--> blackholing {}", partitionedNode);
+                partitionedNode.blackhole();
+            } else {
+                logger.info("--> disconnecting {}", partitionedNode);
+                partitionedNode.disconnect();
+            }
+            cluster.setEmptySeedHostsList();
+            cluster.stabilise();
+
+            partitionedNode.heal();
+            cluster.runRandomly(false, true, EXTREME_DELAY_VARIABILITY);
+            cluster.stabilise();
         }
     }
 }
